Move test to cmd package
